services/anomaly-detector/requirements.txt
fastapi
uvicorn
kafka-python
scikit-learn
joblib
numpy
prometheus-client
